refactored endpoints GetVoteForAddress, getVotesForAddress

+++ backend/main/models/proposal.go
@@ -20,7 +20,7 @@ type Proposal struct {
 	Community_id         int                     `json:"communityId"`
 	Choices              []s.Choice              `json:"choices" validate:"required"`
 	Strategy             *string                 `json:"strategy,omitempty"`
-	Max_weight           *uint64                 `json:"maxWeight,omitempty"`
+	Max_weight           *float64                `json:"maxWeight,omitempty"`
 	Min_balance          *uint64                 `json:"minBalance,omitempty"`
 	Creator_addr         string                  `json:"creatorAddr" validate:"required"`
 	Start_time           time.Time               `json:"startTime" validate:"required"`
